Move DatadogProvider into the SDK directory
