+++ packages/core/src/rum/instrumentation/resourceTracking/implementation/DatadogRumResource/__tests__/ResourceReporter.test.ts
@@ -6,7 +6,7 @@
 
 import { NativeModules } from 'react-native';
 
-import { BufferSingleton } from '../../../../../../DatadogProvider/Buffer/BufferSingleton';
+import { BufferSingleton } from '../../../../../../sdk/DatadogProvider/Buffer/BufferSingleton';
 import type { RUMResource } from '../../../domain/interfaces/RumResource';
 import { ResourceReporter } from '../ResourceReporter';
